# speed up large polygon intersections using STRtrees and many other things