--- a/GeometryOpsCore/src/applyreduce.jl
+++ b/GeometryOpsCore/src/applyreduce.jl
@@ -135,27 +135,6 @@ import Base.Threads: nthreads, @threads, @spawn
     # Map over the chunks
     tasks = map(task_chunks) do chunk
         # Spawn a task to process this chunk
-        Threads.@spawn begin
-            # Where we map `f` over the chunk indices
-            mapreduce(f, op, chunk; init)
-        end
-    end
-
-    # Finally we join the results into a new vector
-    return mapreduce(fetch, op, tasks; init)
-end
-
-@inline function _mapreducetasks(f::F, op, taskrange, threaded::_TrueButStable; init) where F
-    ntasks = length(taskrange)
-    # Customize this as needed.
-    # More tasks have more overhead, but better load balancing
-    tasks_per_thread = 2
-    chunk_size = max(1, ntasks ÷ (tasks_per_thread * nthreads()))
-    # partition the range into chunks
-    task_chunks = Iterators.partition(taskrange, chunk_size)
-    # Map over the chunks
-    tasks = map(task_chunks) do chunk
-        # Spawn a task to process this chunk
         StableTasks.@spawn begin
             # Where we map `f` over the chunk indices
             mapreduce(f, op, chunk; init)