--- a/GeometryOpsCore/src/constants.jl
+++ b/GeometryOpsCore/src/constants.jl
@@ -1,0 +1,8 @@
+"The semi-major axis of the WGS84 ellipsoid"
+const WGS84_EARTH_SEMI_MAJOR_RADIUS = 6378137.0
+
+"The inverse flattening of the WGS84 ellipsoid"
+const WGS84_EARTH_INV_FLATTENING = 298.257223563
+
+"The mean radius of the WGS84 ellipsoid, used for spherical manifold default"
+const WGS84_EARTH_MEAN_RADIUS = 6371008.8