--- a/ext/GeometryOpsLibGEOSExt/segmentize.jl
+++ b/ext/GeometryOpsLibGEOSExt/segmentize.jl
@@ -23,7 +23,7 @@ end
 # 2 behaviours:
 # - enforce: enforce the presence of a kwargs
 # - fetch: fetch the value of a kwargs, or return a default value
-@inline function GO.segmentize(alg::GEOS, geom; threaded::Union{Bool, GO.BoolsAsTypes} = _False())
+@inline function GO.segmentize(alg::GEOS, geom; threaded::Union{Bool, GO.BoolsAsTypes} = False())
     max_distance = enforce(alg, :max_distance, GO.segmentize)
     return GO.apply(
         Base.Fix2(_wrap_and_segmentize_geos, max_distance),