--- a/src/methods/clipping/difference.jl
+++ b/src/methods/clipping/difference.jl
@@ -37,7 +37,7 @@ function difference(
 ) where {T<:AbstractFloat}
     return _difference(
         alg, TraitTarget(target), T, GI.trait(geom_a), geom_a, GI.trait(geom_b), geom_b;
-        exact = _True(), kwargs...,
+        exact = True(), kwargs...,
     )
 end