… .github/workflows/CI.yml
… Project.toml
… benchmarks/benchmarks.jl
… docs/Project.toml
… docs/src/explanations/peculiarities.md
… src/methods/clipping/coverage.jl
… src/methods/clipping/cut.jl
… src/methods/clipping/intersection.jl
… src/methods/clipping/predicates.jl
… src/methods/clipping/union.jl
… src/methods/geom_relations/coveredby.jl
… src/methods/geom_relations/crosses.jl
… src/methods/geom_relations/disjoint.jl
… src/methods/geom_relations/geom_geom_processors.jl
… src/methods/geom_relations/overlaps.jl
… src/methods/geom_relations/touches.jl
… src/methods/geom_relations/within.jl
… src/transformations/segmentize.jl
… src/types.jl
… src/utils/LoopStateMachine.jl
… src/utils/LoopStateMachine/LoopStateMachine.jl
… src/utils/NaturalIndexing.jl
… src/utils/STRDualQuery.jl
… src/utils/SpatialTreeInterface/NaturalIndex.jl
… src/utils/SpatialTreeInterface/README.md
… src/utils/SpatialTreeInterface/STRDualQuery.jl
… src/utils/SpatialTreeInterface/SpatialTreeInterface.jl
… src/utils/utils.jl
… test/external/jts/jts_testset_reader.jl
… test/external/tg/data/LICENSE.md