> [!NOTE]
> This is effectively the GeometryOps dev branch now.

\This PR uses STRtrees to substantially accelerate polygon queries.   It pulls the actual looping all the way out and allows the use of STRtrees to skip substantial sections of polygons.

There are three methods here:
- naive, n*m nested loop
- loop over all edges of a, strtree on b
- dual tree query on strtrees from a and b


This algorithm should be easily portable to s2 when that becomes available - and the structure allows e.g. manifold passthrough as well.
<img width=400 src="https://github.com/user-attachments/assets/cfd4fcf0-2a22-4504-9420-9ef1cd384a6b" />


***

This PR also adds a LoopStateMachine module that basically just defines an `Action` wrapper struct, and a macro that can take care of that action wrapper struct.

With this macro, a function running within a loop can return `Break()` or `Continue()` which indicates to the caller that it should break or continue.

E.g.:
```julia
f(i) = i == 3 ? Break() : i
count = 1
for i in 1:5
    count = @processloopaction f(i)
end
count # 2
```
    

***

TODOs: 
- [x] add rejection tests for duplicate points
- [x] add lopsided single tree query
- [x] add dual tree query
- [x] figure out good heuristics for when to choose which
- [ ] add more tests and benchmarks with complex polygons