@asinghvi17 2025-02-23
I just added the acceleration to `_intersection_points` as well.  I will need some help accelerating the predicates since they seem to have an outer loop that is affected by the result of the inner loop, but can try to do some acceleration in the RectTrait case.