@rafaqz 2025-02-23
Can you link to the predicates code with the loop?