@asinghvi17 2025-02-23
https://github.com/JuliaGeo/GeometryOps.jl/blob/main/src%2Fmethods%2Fgeom_relations%2Fgeom_geom_processors.jl#L121-L208