@rafaqz 2025-02-23
Right. There is already a _maybe_skip_disjoint check further out too: https://github.com/JuliaGeo/GeometryOps.jl/blob/main/src%2Fmethods%2Fgeom_relations%2Fgeom_geom_processors.jl#L113