@rafaqz 2025-02-23
Ok I see what you mean, the al is generating edges on the fly. 
Probably the alg can be changed to generate a vector of edges? 

We can reuse the memory... And then you can STRtree them