@asinghvi17 2025-02-23
Yeah...the alg can also increment the value of `i` here:

https://github.com/JuliaGeo/GeometryOps.jl/blob/18eb5e9aee760b0b6a190ea12c3c17b00c7a6309/src/methods/geom_relations/geom_geom_processors.jl#L160