@rafaqz 2025-02-23
Can the tree be for j? 

Looks like it's not straightforward either way but seems doable