@asinghvi17 2025-02-23
Maybe yeah, that would at least save us querying intersection_point all the time