@asinghvi17 2025-02-28
The last 4 commits before this comment implement the natural tree from `tg` as well as a spatial tree interface that should be valid for any spatial tree.   The functions that SpatialTreeInterface enables are sufficiently generic that they _should_ be able to work for any number of dimensions and potentially even on the sphere.

The natural tree has been tested to perform a scanline of Boston's coordinates in the USA polygon, 35000 vertices, in ~80 ns!  That's insane timing.  It ought to speed up the point in polygon to be similar in speed to the C library one, if we don't use exact predicates.