@asinghvi17 2025-02-28
```
# Original

julia> @be GO.contains($(usa_poly), $((-76, 42)))
Benchmark: 3590 samples with 4 evaluations
 min    5.729 μs
 median 6.604 μs
 mean   6.426 μs
 max    16.208 μs

# After ExactPredicates
@be GO.contains($(usa_poly), $((-76, 42)))
Benchmark: 3672 samples with 4 evaluations
 min    5.708 μs
 median 6.396 μs
 mean   6.235 μs
 max    8.198 μs

# After NaturalIndex (this IS constructing the natural index as well)
# this is substantially slower than brute force!
julia> @be GO.contains($(usa_poly), $((-76, 42)))
Benchmark: 1359 samples with 1 evaluation
 min    22.291 μs (11 allocs: 403.688 KiB) # compared to 14 microseconds for TG given a GI geom
 median 38.500 μs (11 allocs: 403.688 KiB) # compared to 38 microseconds for TG given a GI geom
 mean   58.483 μs (11 allocs: 403.688 KiB, 1.34% gc time)
 max    3.341 ms (11 allocs: 403.688 KiB, 98.75% gc time)

# With pre-computed natural index on usa_poly
# 2.8 OOM faster!
julia> @be GO.contains($(prepare_naturally(usa_poly)), $((-76, 42)))
Benchmark: 3629 samples with 66 evaluations
 min    341.545 ns (1 allocs: 16 bytes)
 median 381.318 ns (1 allocs: 16 bytes)
 mean   390.878 ns (1 allocs: 16 bytes)
 max    901.515 ns (1 allocs: 16 bytes)
```