@rafaqz 2025-02-28
So fast!!

 I guess we can reduce those allocs a bit or something