@asinghvi17 2025-03-01
Turns out intersection_point tests are failing because the algorithm is returning duplicate intersection points.