@asinghvi17 2025-03-01
Brief experiment to confirm whether a lazy generator edge list has equivalent speed to a non-lazy version:

```julia
# Setup code
import GeometryOps as GO
import GeometryOps.SpatialTreeInterface as STI
using Extents

function doityourself(f, pred, geom)
       p1 = GI.getpoint(geom, 1)
       #p2 = GI.getpoint(geom, 2)
       for i in 2:GI.npoint(geom)
           p2 = GI.getpoint(geom, i)
           ext = Extent(X = extrema(GI.x, (p1, p2)), Y = extrema(GI.y, (p1, p2)))
           pred(ext) && f(i)
           p1 = p2
       end
 end


# get data together
using NaturalEarth

all_countries = naturalearth("admin_0_countries", 10)^C
usa_multipoly = all_countries.geometry[findfirst(==("United States of America"), all_countries.NAME)] |> GO.tuples #x -> GI.convert(LG, x) |> LG.makeValid |> GO.tuples^C
usa_poly = GI.getgeom(usa_multipoly, findmax(GO.area.(GI.getgeom(usa_multipoly)))[2]) # isolate the poly with the most area^C
usa_ring = GI.getexterior(usa_poly)


# now try things
using Chairmarks
# benchmark: find all edges that cross 42 degrees latitude
tree_lazy = STI.FlatNoTree(GO.lazy_edge_extents(usa_ring))
tree_eager = STI.FlatNoTree(GO.edge_extents(usa_ring))
tree_natural = GO.NaturalIndexing.NaturalIndex(GO.lazy_edge_extents(usa_ring))
```
Natural indexing (fastest, of course)
```julia-repl
julia> @be Int[] STI.do_query(Base.Fix1(push!, _), $(ext -> ext.Y[1] <= 42 <= ext.Y[2]), $tree_natural)
Benchmark: 2717 samples with 126 evaluations
 min    205.357 ns (0.06 allocs: 177.524 bytes)
 median 224.532 ns (0.06 allocs: 177.524 bytes)
 mean   276.546 ns (0.06 allocs: 177.524 bytes, 0.11% gc time)
 max    36.939 μs (0.06 allocs: 177.524 bytes, 98.80% gc time)
```
Lazy extent list in FlatNoTree (what it says on the tin - no tree at all)
```julia-repl
julia> @be Int[] STI.do_query(Base.Fix1(push!, _), $(ext -> ext.Y[1] <= 42 <= ext.Y[2]), $tree_lazy)
Benchmark: 4073 samples with 2 evaluations
 min    11.188 μs (1 allocs: 232 bytes)
 median 11.229 μs (1 allocs: 232 bytes)
 mean   11.490 μs (1 allocs: 232 bytes)
 max    48.896 μs (1 allocs: 232 bytes)
```
Materialized vector of extents in eager tree
```julia-repl
julia> @be Int[] STI.do_query(Base.Fix1(push!, _), $(ext -> ext.Y[1] <= 42 <= ext.Y[2]), $tree_eager)
Benchmark: 4983 samples with 3 evaluations
 min    5.611 μs (0.67 allocs: 154.667 bytes)
 median 6.194 μs (0.67 allocs: 154.667 bytes)
 mean   6.200 μs (0.67 allocs: 154.667 bytes)
 max    15.597 μs (0.67 allocs: 154.667 bytes)
```
DIY solution - compute extent and check on the fly
```julia-repl
julia> @be Int[] doityourself(Base.Fix1(push!, _), $(ext -> ext.Y[1] <= 42 <= ext.Y[2]), $(usa_ring))
Benchmark: 4080 samples with 2 evaluations
 min    11.146 μs (1 allocs: 232 bytes)
 median 11.188 μs (1 allocs: 232 bytes)
 mean   11.332 μs (1 allocs: 232 bytes)
 max    21.354 μs (1 allocs: 232 bytes)
```


It's clear that the lazy indexing takes the same amount of time as the DIY solution, meaning we can actually